+++ tests/test_commands_blender_render.py
@@ -1,3 +1,6 @@
+from pathlib import Path
+import subprocess
+
 from unittest.mock import patch, call
 
 from test_runner import AbstractCommandTest
@@ -67,8 +70,6 @@ class BlenderRenderTest(AbstractCommandTest):
 
     def test_cli_args(self):
         """Test that CLI arguments in the blender_cmd setting are handled properly."""
-        from pathlib import Path
-        import subprocess
         from mock_responses import CoroMock
 
         filepath = str(Path(__file__).parent)
@@ -94,6 +95,41 @@ class BlenderRenderTest(AbstractCommandTest):
                 '-noaudio',
                 '--background',
                 filepath,
+                '--render-format', 'JPEG',
+                '--render-frame', '1..2',
+                stdin=subprocess.DEVNULL,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.STDOUT,
+            )
+
+    def test_python_expr(self):
+        from mock_responses import CoroMock
+
+        filepath = str(Path(__file__).parent)
+        settings = {
+            # Point blender_cmd to this file so that we're sure it exists.
+            'blender_cmd': '%s --with --cli="args for CLI"' % __file__,
+            'python_expr': 'print("yay in \'quotes\'")',
+            'chunk_size': 100,
+            'frames': '1..2',
+            'format': 'JPEG',
+            'filepath': filepath,
+        }
+
+        cse = CoroMock(...)
+        cse.coro.return_value.wait = CoroMock(return_value=0)
+        with patch('asyncio.create_subprocess_exec', new=cse) as mock_cse:
+            self.loop.run_until_complete(self.cmd.run(settings))
+
+            mock_cse.assert_called_once_with(
+                __file__,
+                '--with',
+                '--cli=args for CLI',
+                '--enable-autoexec',
+                '-noaudio',
+                '--background',
+                filepath,
+                '--python-expr', 'print("yay in \'quotes\'")',
                 '--render-format', 'JPEG',
                 '--render-frame', '1..2',
                 stdin=subprocess.DEVNULL,